# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Credit Card Notifications and Alerts API
  version: 1.0.0
servers:
- url: http://localhost:3000
paths:
...
  /balance-alert:
    post:
      operationId: balanceAlert
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/balanceAlert_request'
        required: true
      responses:
        "200":
          content:
            text/plain:
              schema:
                type: string
          description: OK
      x-content-type: application/json
      x-accepts: text/plain
  /collection-notification:
    post:
      operationId: collectionNotification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/collectionNotification_request'
        required: true
      responses:
        "200":
          content:
            text/plain:
              schema:
                type: string
          description: OK
      x-content-type: application/json
      x-accepts: text/plain
...
components:
  schemas:
...
    balanceAlert_request:
      properties:
        paymentDueDate:
          format: date
          type: string
        currentDate:
          format: date
          type: string
        cardLast4:
          maxLength: 4
          minLength: 4
          type: string
      required:
      - cardLast4
      - currentDate
      - paymentDueDate
      type: object
    collectionNotification_request:
      properties:
        delinquencyStatus:
          type: string
        outstandingBalance:
          type: string
        additionalCharges:
          type: string
        cardLast4:
          maxLength: 4
          minLength: 4
          type: string
      required:
      - additionalCharges
      - cardLast4
      - delinquencyStatus
      - outstandingBalance
      type: object
...